% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champagne.R
\name{f}
\alias{f}
\title{Champagne glass function (vectorized, using indicator functions and `structure()`)}
\usage{
f(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector.
}
\description{
A vectorized version of the Champagne glass function using indicator functions and `structure()`.
}
\examples{
f(0)
f(0.5)
f(10)
f(15)
f(20)
f(21)
}
\author{
samorso
}
