% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champagne.R
\name{f1}
\alias{f1}
\title{Champagne glass function (non-vectorized)}
\usage{
f1(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector.
}
\description{
A non-vectorized version of the Champagne glass function.
}
\examples{
f1(0)
f1(0.5)
f1(10)
f1(15)
f1(20)
f1(21)
}
\author{
samorso
}
