% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champagne.R
\name{f2}
\alias{f2}
\title{Champagne glass function (vectorized, using `purrr::map_dbl()`)}
\usage{
f2(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector.
}
\description{
A vectorized version of the Champagne glass function using `purrr::map_dbl()`.
}
\examples{
f2(0)
f2(0.5)
f2(10)
f2(15)
f2(20)
f2(21)
}
\author{
samorso
}
