% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champagne.R
\name{f3}
\alias{f3}
\title{Champagne glass function (vectorized, using `vapply()`)}
\usage{
f3(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector.
}
\description{
A vectorized version of the Champagne glass function using `vapply()`.
}
\examples{
f3(0)
f3(0.5)
f3(10)
f3(15)
f3(20)
f3(21)
}
\author{
samorso
}
