% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champagne.R
\name{f4}
\alias{f4}
\title{Champagne glass function (vectorized, using `Vectorize()`)}
\usage{
f4(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector.
}
\description{
A vectorized version of the Champagne glass function using `Vectorize()`.
}
\examples{
f4(0)
f4(0.5)
f4(10)
f4(15)
f4(20)
f4(21)
}
\author{
samorso
}
