% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champagne.R
\name{f5}
\alias{f5}
\title{Champagne glass function (vectorized, using indicator functions)}
\usage{
f5(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector.
}
\description{
A vectorized version of the Champagne glass function using indicator functions.
}
\examples{
f5(0)
f5(0.5)
f5(10)
f5(15)
f5(20)
f5(21)
}
\author{
samorso
}
