% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champagne.R
\name{g}
\alias{g}
\title{Champagne glass function}
\usage{
g(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector.
}
\description{
A function that represents the side cut of half of the Champagne glass placed horizontally on an imaginary x-axis as the following piecewise function.
\deqn{f(x) = \begin{cases} 15 & 0 \leq x < 0.5 \\ 2 & 0.5 \leq x < 10 \\ 8 \log_2(|x - 9|) + 2 & 10 \leq x < 15 \\ 8 \log_2(6) + 2 & 15 \leq x \leq 20 \\ 0 & \text{otherwise} \end{cases}}
}
\examples{
champagne(0)
champagne(0.5)
champagne(10)
champagne(15)
champagne(20)
champagne(21)
}
\author{
samorso
}
